bv: Replace constructors from integers with static functions.
There are issues with different compiler versions on different platforms
with ambiguous overloads for the constructors
  `BitVector(uint64_t size, uint64_t value)` and
  `BitVector(uint64_t size, int64_t value)`.
This commit replaces these constructors with static functions
  `from_ui(uint64_t size, uint64_t value` and
  `from_si(uint64_t size, int64_t value)`.
We maintain an explicit separation of the two and do not keep the (more
often used) constructor from uint64_t values to make the distinction
very explicit. For negative signed values, using the wrong
constructor/construction method is dangerous since the value has to be
sign extended for bit-widths > 64 (but the construction from uint64_t
values zero extends it).